vcs: Added a vcs module, with Git and Subversion support.
This provides a mechanism to interact with VCS-controlled projects.  It
currently supports Git and Subversion.  The main interesting feature
here is a rule to generate a version string.  There are other rules for
querying a directory for its properties that may also prove generally
useful.

There are also rules to 'fetch' from a repository into a directory and
'checkout' a symbolic reference from a directory.  These could be used
to forcibly get a specific revision of a package from a repository.  It
could also be used to create a proper source package management system
within Boost.Build.

There are two examples to show how this module might be used.

The documentation build has not been tested.  However, a
reStructuredText document that should be identical to the BoostBook file
is included in the doc directory for reference.

There are no tests, but there are two examples that do quite a bit of
work to test the module, but do not verify the results.

This has been tested on Linux, OS X, and Windows.  The 'type' rule fails
on Windows due to `path.exists' rule failing.  This needs to be
investigated.

This does not implement these functions in Python.  It should be
straightforward to do so.