vcs: Improve documentation.
This change fixes some bugs in the documentation.  It also improves the
documentation generated by Boost.Build.  The generated documentation is
now much closer to the BoostBook and reStructuredText documentation.

The fact that there are at minimum two versions of the documentation
that is required by Boost.Build is a bit of a bummer.